--- src/components/fullWidthlabel.js
@@ -1,4 +1,0 @@
-import { Label } from 'reactstrap';
-
-export const FullWidthLabelField = props => <Label {...props} sm="12" />;
-export const FullWidthLabelTextarea = props => <Label {...props} sm="5" />;
