+++ src/components/labels.js
@@ -1,0 +1,4 @@
+import { Label } from 'reactstrap';
+
+export const LargeWidthLabelField = props => <Label {...props} sm="12" />;
+export const MediumWidthLabelTextarea = props => <Label {...props} sm="5" />;
